DB/SAI: Disclosure (12710)
Fixes #7192

The credit timing is very strange, it is given right at the start of the script, but that's how it is from Blizzard, so we'll stick to that.

As for the areatriggers, I know I could've just simplified it and made it into an areatrigger_teleport. But the sniffs state that it is the trigger that just makes the bunnies cast the teleport spell, so we'll keep that one complicated.
This is the penultimate quest of the chain, the final quest is a WIP. Any additional info would be very helpful.

Regards, Kirk.